+++ shuup/core/models/_suppliers.py
@@ -12,6 +12,7 @@ from django.utils.translation import ugettext_lazy as _
 from enumfields import Enum, EnumIntegerField
 from filer.fields.image import FilerImageField
 from jsonfield import JSONField
+from timezone_field.fields import TimeZoneField
 from parler.managers import TranslatableQuerySet
 from parler.models import TranslatedFields
 
@@ -82,6 +83,10 @@ class Supplier(ModuleInterface, TranslatableShuupModel):
     )
     is_approved = models.BooleanField(default=True, verbose_name=_("approved"), help_text=_(
         "Indicates whether this supplier is currently approved."
+    ))
+    timezone = TimeZoneField(blank=True, null=True, verbose_name=_('time zone'), help_text=_(
+        "The timezone in which the shop resides. This can be used to target the delivery of promotional materials "
+        "at a particular time."
     ))
     options = JSONField(blank=True, null=True, verbose_name=_("options"))
     translations = TranslatedFields(